requirements.txt
ace_tools
annotated-types
antlr4-python3-runtime
anyio
asttokens
av
certifi
cffi
charset-normalizer
click
colorama
coloredlogs
comm
contourpy
ctranslate2
cycler
debugpy
decorator
executing
fastapi
faster-whisper
filelock
flatbuffers
fonttools
fsspec
h11
huggingface-hub
humanfriendly
idna
ipykernel
ipython
jedi
Jinja2
joblib
jupyter_client
jupyter_core
kiwisolver
MarkupSafe
matplotlib
matplotlib-inline
mpmath
nest-asyncio
networkx
numpy
omegaconf
onnxruntime
packaging
pandas
parso
pillow
platformdirs
prettytable
prompt_toolkit
protobuf
psutil
pure_eval
pyasn1
pycparser
pydantic
pydantic_core
Pygments
pyparsing
pyreadline3
python-dateutil
pytz
PyYAML
pyzmq
regex
requests
rsa
safetensors
scikit-learn
scipy
sentencepiece
silero-vad
six
sniffio
sounddevice
soundfile
stack-data
starlette
sympy
tabulate
threadpoolctl
tokenizers
torch
torchaudio
tornado
tqdm
traitlets
transformers
typing-inspection
typing_extensions
tzdata
urllib3
uvicorn
wcwidth
websockets
python-dotenv
crewai
bs4
faiss-cpu